refactor: reorganize notify inbox views and update imports
