+++ invenio_notify/views/__init__.py
@@ -1,0 +1,3 @@
+
+# in order to make `invenio_base.api_blueprints`, api endpoints must be imported
+from invenio_notify.views.api_views import *  # noqa
