+++ invenio_notify/views/api_views.py
@@ -1,8 +1,4 @@
 from flask import current_app, request, jsonify
-from invenio_administration.views.base import (
-    AdminResourceListView, AdminResourceDetailView,
-)
-from invenio_i18n import lazy_gettext as _
 from invenio_oauth2server import require_oauth_scopes, require_api_auth
 from invenio_pidstore.errors import PIDDoesNotExistError
 
@@ -11,70 +7,6 @@ from invenio_notify import constants
 from invenio_notify.blueprints import rest_blueprint
 from invenio_notify.scopes import inbox_scope
 from invenio_notify.services.service import NotifyInboxService
-
-
-class NotifyInboxListView(AdminResourceListView):
-    """Search admin view."""
-
-    api_endpoint = "/notify-inbox"
-    name = "notify-inbox"
-    resource_config = "notify_inbox_resource"
-    title = "Notify Inbox"
-    menu_label = "Notify Inbox"
-    category = _("Notify Inbox")
-    pid_path = "id"
-    icon = "newspaper"
-
-    display_search = True
-    display_delete = True
-    display_create = False
-    display_edit = False
-
-    item_field_list = {
-        "id": {"text": _("Id"), "order": 1, "width": 1},
-        "raw": {"text": _("Raw"), "order": 2, "width": 7},
-        "recid": {"text": _("Record ID"), "order": 3, "width": 1},
-        "user_id": {"text": _("User ID"), "order": 4, "width": 1},
-        "process_date": {"text": _("Process Date"), "order": 5, "width": 2},
-        "process_note": {"text": _("Process Note"), "order": 6, "width": 2},
-        "created": {"text": _("Created"), "order": 7, "width": 2},
-        "updated": {"text": _("Updated"), "order": 8, "width": 2},
-    }
-
-    create_view_name = None
-
-    search_config_name = "NOTIFY_SEARCH"
-
-    search_sort_config_name = "NOTIFY_SORT_OPTIONS"
-
-
-class NotifyInboxDetailView(AdminResourceDetailView):
-    """Admin notify inbox detail view."""
-
-    url = "/notify-inbox/<pid_value>"
-    api_endpoint = "/notify-inbox"
-    name = "notify-inbox-details"
-    resource_config = "notify_inbox_resource"
-    title = "Notify Inbox Details"
-
-    display_delete = True
-    display_edit = False
-
-    list_view_name = "notify-inbox"
-    pid_path = "id"
-
-    item_field_list = {
-        "id": {"text": _("Id"), "order": 1, "width": 1},
-        "raw": {"text": _("Raw"), "order": 2, "width": 7},
-        "recid": {"text": _("Record ID"), "order": 3, "width": 1},
-        "user_id": {"text": _("User ID"), "order": 4, "width": 1},
-        "process_date": {"text": _("Process Date"), "order": 5, "width": 2},
-        "process_note": {"text": _("Process Note"), "order": 6, "width": 2},
-        "created": {"text": _("Created"), "order": 7, "width": 2},
-        "updated": {"text": _("Updated"), "order": 8, "width": 2},
-    }
-
-
 
 
 @rest_blueprint.route("/inbox", methods=['POST'])
